For some reason all of these commits didn't get out to the main branch:
Adding in vendor js.
Adding in css for the view.
Adding in the general html view with a switching view for the Entity and Analysis Views on the toolbar.

Adding in js framework for our code with classes.

Adding in data loading. All documents and entities are now loaded into classes and the objects are linked (e.g. a Entity has a reference of all the documents it is contained in).
IMPORTANT: To get access to an entity or document use: "data.documents" or "data.entities" they both return arrays.

Still need to add a time array and add the actual text of each document, but this is a start.